+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 
 -   GP2-1068 - adopt Black auto-formatting + provide optional pre-commit config
+-   No ticket - added Activity Stream endpoint to list users
 
 ### Fixed bugs: